Rename variables after network module changes
The network module now creates public and private subnets. By default, both master and worker nodes use all the subnets in the common "subnet_ids" list. But it is possible to specify a different list for worker nodes in "subnet_ids_worker". This is useful to restrict the worker nodes to a single AZ, for example