Add spot_instance_pools argument to advanced node groups
The spot instance pools argument must be set to 0 when "capacity-optimized" is used as the spot allocation strategy.

Sometimes when we do not explicitly define it to 0, it is set to 2 by the terraform AWS provider, and there is an error: "SpotInstancePools option is only available with the lowest-price allocation strategy".

+ update the node group examples in the README, after several modifications in the node group API.