Disable EKS-defined max number of pods per node
Add a variable "node_eks_max_pods".

When setting this variable to false, AWS-EKS will not set a max number of pods that is based on the instance type, in the Kubelet config.

This is useful when using the Calico CNI to overcome the IP addresses limitations.

This changes only the launch template: the existing nodes will not be affected, nodes will need to be roll-updated.

Limitation: only the node groups non-managed by EKS ("node_groups_advanced") will
benefit from this setting. Currently there is no way to disable max-pods
with managed node groups.