Fix: add custom tags on managed node's ASG (fixes scale-up by the autoscaler)
The tags specified on the resource type "aws_eks_node_group" are not propagated to the ASG that represents this node group (issue https://github.com/aws/containers-roadmap/issues/608).

As a workaround, we add tags to the ASG after the nodegroup creation/updates using the AWS command-line.

This will fix scaling up from 0, in EKS-managed node groups, when pods
have affinities/nodeSelectors defined on custom tags.